requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
plotly>=5.20.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
pgeocode>=0.4.0
prophet>=1.1.5
openai>=1.30.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.29.0
openpyxl>=3.1.0
